xen: add a simple buffered reader on top of the raw Console
This allows us to read the exact number of bytes that we want, and
cache any extra we might receive (since the backend xentropyd might
be more generous than we need).

This also allows us to read the handshake message even if it is
returned in the same buffer as the first chunk of entropy.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>