Solo5 support
There are two opam files and two META files, but OCaml and C code is the same.
Depending on the target, either mirage-solo5 or mirage-xen is dragged in, and
appropriate CFLAGS (using pkg-config ocaml-freestanding or mirage-xen) are
passed through.

There is no fancy copying around needed, and no hacks in the _tags file