mirage-crypto-entropy: adapt entropy feeding to how other OS (FreeBSD) do it: - in the interrupt_hook (called on every Lwt loop enter), collect some bits from __rdtsc - in a separate task, call rdrand periodically (each second) for each pool
the choice to use rdrand instead of rdseed is due to the fact that rdseed is
roughly 3 times as expensive and the CPU can run out of it. since we're mixing
the output into the fortuna pools anyways (and not use the output as only seed
of the RNG), this should be safe.

the previous behaviour was, especially under high load, a bit problematic, since
each time the event loop was triggered, rdseed got invoked which is pretty
expensive.