Entropy source registration
Instead of a closed variant, provide `register_source : string -> source`,
and use the type `source` for accumulation etc. This has the advantage of
being flexible (if some other library wants to contribute to the entropy
collection), and the only function on the variant was to extract a list of
available sources, plus a pretty printer. This is still available (now as
`Entropy.pp_sources : unit Fmt.t`), while the list of entropy sources is
not anymore.

fixes #68