mirage-crypto-rng: require client code to match on cpu_rng support
To avoid scheduling timers that do no work, require client code to match
on the presence of cpu_rng support before attempting to use it.

It appears that cpu_rng_bootstrap was kindof already doing this, but
with an exception. In this cases, this PR just makes what's going on a
little more explicit.