bswap32/64: avoid __builtin_bswap since this is optimized by the C compiler (with -O3 -mssse3) to the SSSE3 PSHUFB instruction (at compile time), which is unavailable on CPUs without SSSE3 (such as AMD Phenom II X4).
Instead, use bswap inline assembly - this is available on CPUs and not optimized
to an SSSE3 insruction. The peformance difference (with bench/speed.exe sha256)
is negligible. This should bring us back to "compile on any CPU, run anywhere"
(optimized code is used if at runtime the respective CPU features (SSSE3,
PCLMULQ, AESNI) are present - the optimized code is always generated (on X86)).

This fixes #93.