mirage-crypto-ec: avoid ~rng in gen_key / generate, provide Dh.secret_of_cs
Previously, private key generation (using a RNG) and decoding a private key from
a buffer used a single function (gen_key / generate). Since ECDH/ECDSA on
NIST curves has some requirements on the values, this potentially executed the
RNG repeatedly. This is fine if an actual RNG is used, but not if a static
buffer was used (which returns the same value) -- which case lead to divergence
(an infinite loop without progress).

Now, the signature of Mirage_crypto_ec.Dsa and Mirage_crypto_pk.Dsa are more
congruent, and Mirage_crypto_ec.Dh and Mirage_crypto_pk.Dh are as well more
congruent. Similarly to Mirage_crypto_pk, the generate functions take a
generator and not a RNG function.