mirage-crypto-rng: on arm32 check whether the performance counter are available
As reported and investigated by @adams-1979 in #113, reading the virtual tick
count on arm32 is only accessible in kernel mode (or if enabled via a kernel
module). Use nsec from clock_gettime as a fallback.

If virtualized via ocaml-freestanding, use the existing code (since Xen masks
(or used to mask) the performance counters). This should be revisited once we
target arm32 as MirageOS platform.