Mirage_crypto_rng_mirage: always feed the default generator
This applies a8c7bbd2552a9d2177450e95f280342f80fba01d (May 2020) to the mirage
rng feeding.

The reasoning behind this change is twofold:
- consistency of the semantics across implementations (lwt, mirage, async);
- make it hard to produce bad random.

Now, the expectation is that a single RNG is used in an application, and this
is fed fresh entropy on a periodic basis. Since the API allows to mutate the
default generator (Mirage_crypto_rng.set_default_generator), which is used
by default by all applications, it is crucial that this one receives fresh
entropy.

The "initialize" function does the following: fail if the default rng is already
set; collect some bootstrap entropy; set the default rng; start tasks that
periodically (every m nanoseconds) collect entropy and feed it to the generator;
register a hook to the scheduler to collect a bit more entropy on every
scheduling decision.