Add a missing dependency in the configuration detection
The configuration detection executable uses `dune-configurator` and so
it uses the configuration passed in (currently) `.dune/configurator.v2`
in particular to know which C compiler it should invoke. This adds an
explicit dependency to this file (with a glob in case the format
changes) to make sure a change of compiler triggers a redetection.